Ashoat's revisions
